# microservices/movie-server/src/main/java/ch/heigvd/amt/movie/api/endpoints/UsersApiController.java
package ch.heigvd.amt.movie.api.endpoints;

import ch.heigvd.amt.movie.api.model.User;
import ch.heigvd.amt.movie.entities.UserEntity;
import ch.heigvd.amt.movie.repositories.UserRepository;
import io.swagger.annotations.Api;
import io.swagger.annotations.ApiParam;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.servlet.support.ServletUriComponentsBuilder;

import javax.validation.Valid;
import java.net.URI;
import java.util.ArrayList;
import java.util.List;
import java.util.Optional;

/**
 * Spring controller that implements all operations generated by codegen
 * @author Gilliand Loris
 * @author Tutic Mateo
 * @version 1.0
 * @since 11.12.2019
 */
@Controller
@Api(tags = "users")
public class UsersApiController implements UsersApi {
    @Autowired
    UserRepository userRepository;

    public ResponseEntity<List<User>> getUsers() {
        List<User> users = new ArrayList<>();
        for (UserEntity userEntity : userRepository.findAll()) {
            users.add(toUser(userEntity));
        }
        return ResponseEntity.ok(users);
    }

    public ResponseEntity<User> getUser(Integer id) {
        Optional<UserEntity> userEntity = userRepository.findById(Long.valueOf(id));
        if (userEntity.isPresent()) {
            User user = toUser(userEntity.get());
            return ResponseEntity.ok(user);
        } else {
            return ResponseEntity.notFound().build();
        }
    }

    public ResponseEntity<Object> createUser(@ApiParam(value = "", required = true) @Valid @RequestBody User user) {
        UserEntity newUserEntity = toUserEntity(user);
        userRepository.save(newUserEntity);

        URI location = ServletUriComponentsBuilder
                .fromCurrentRequest().path("/{id}")
                .buildAndExpand(newUserEntity.getId()).toUri();

        return ResponseEntity.created(location).build();
    }

    public ResponseEntity<Void> deleteUser(Integer id) {
        if (userRepository.findById(Long.valueOf(id)).isPresent()) {
            userRepository.deleteById(Long.valueOf(id));
            return ResponseEntity.status(204).build();
        } else {
            return ResponseEntity.notFound().build();
        }
    }

    public ResponseEntity<Void> updateUser(Integer id, @Valid @RequestBody User user) {
        UserEntity userEntity = toUserEntity(user);
        if (userRepository.findById(Long.valueOf(id)).isPresent()) {
            userEntity.setId(id);
            userRepository.save(userEntity);
            return ResponseEntity.ok().build();
        } else {
            return ResponseEntity.notFound().build();
        }
    }

    private UserEntity toUserEntity(User user) {
        UserEntity entity = new UserEntity();
        entity.setLastname(user.getLastname());
        entity.setFirstname(user.getFirstname());
        return entity;
    }

    private User toUser(UserEntity entity) {
        User user = new User();
        user.setLastname(entity.getLastname());
        user.setFirstname(entity.getFirstname());
        return user;
    }
}